Updated database-tests
