+++ test/scripts/database/async/extended/index.js
@@ -1,0 +1,10 @@
+require('./data');
+require('./length');
+
+require('./contains');
+require('./generateId');
+require('./get');
+require('./getNormal');
+require('./set');
+require('./setDefault');
+require('./update');
