+++ test/scripts/database/async/index.js
@@ -1,0 +1,7 @@
+// jshint esversion: 8
+
+describe('Database', () => {
+
+  require('./extended');
+
+});
