+++ test/scripts/database/index.js
@@ -1,0 +1,7 @@
+// jshint esversion: 8
+
+describe('database', () => {
+
+  require('./async');
+
+});
