+++ test/scripts/index.js
@@ -1,2 +1,3 @@
 require('./adapters');
+require('./database');
 require('./database_object');
